data_utility/pyproject.toml
[tool.poetry]
name = "data_utility"
version = "0.1.0"
description = "Provides utility for working with several ASR and SV datasets"
authors = ["Nik Vaessen <git@mail.vaessen.tech>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
torchdata = "^0.5.1"
black = "^22.8.0"
click = "^8.1.3"
torch = "^1.13.1"
torchaudio = "^0.13.1"
numpy = "^1.23.2"
jiwer = "^2.4.0"
SoundFile = "^0.10.3.post1"
tqdm = "^4.64.1"
yaspin = "^2.2.0"
pandas = "^1.4.4"
pandera = "^0.12.0"
webdataset = "^0.2.20"
scipy = "^1.9.1"
scikit-learn = "^1.1.2"
jupyterlab = "^3.4.7"
transformers = "^4.22.2"
datasets = "^2.5.1"
seaborn = "^0.12.0"
ipywidgets = "^8.0.2"
ffmpeg-python = "^0.2.0"
more-itertools = "^9.0.0"
openai-whisper = "^20230117"
langcodes = "^3.3.0"

[tool.poetry.scripts]
verify_checksum = 'data_utility.scripts.verify_checksum:main'
convert_to_wav = 'data_utility.scripts.convert_to_wav:main'
write_tar_shards = 'data_utility.scripts.write_tar_shards:main'
collect_statistics = 'data_utility.scripts.collect_statistics:main'
generate_character_vocabulary = 'data_utility.scripts.generate_character_vocabulary:main'
generate_character_distribution= 'data_utility.scripts.generate_character_distribution:main'
generate_speaker_mapping = 'data_utility.scripts.generate_speaker_mapping:main'
generate_speaker_trials = 'data_utility.scripts.generate_speaker_trials:main'
transcribe_whisper = 'data_utility.scripts.transcribe_whisper:main'
split_csv = "data_utility.scripts.split_csv:main"
# specific for librispeech
ls_generate_csv = 'sets.librispeech.generate_csv:main'
# specific for voxceleb
vc_generate_csv = 'sets.voxceleb.generate_csv:main'
# specific for switchboard
sw_generate_csv = 'sets.switchboard.generate_csv:main'
sw_to_utterance = 'sets.switchboard.switchboard_to_utterances:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
